Fix number with range slider preference list item
The number-with-range-sliders had a number of subtle issues with
showing and storing the correct input while the user is interacting
with it. This change allows the display value to be the empty string
or just a period while the user is interacting with it.
This also makes sure that when a user puts in a number with a specific number
of digits, we show all the digits even when it's beyond the specificity
indicated by the "digits" property passed to the component.