Refactor preference list and update styles
Pass the control object to PreferenceListItem as a prop so that
the PreferenceListItem does not have to have several props that
it only needs so that it can pass thru to one of its children,
the control.

Rewrite styles for most preference list items so that everything
lines up uniformly and the scales of things are under our "direct"
control. For some styles, there were subtle difference between the
rules I wanted to apply on this screen and the ones that were in
shared.scss, so I wrote a few new styles in shared.scss that look
like mostly duplicates. This can be addressed better in a followup
PR. It is easier to check this in now in a working state.

Styling on mobile still needs fixing. The react components can still
be made cleaner and clearer.