Add warning for preferences that require refresh
Some preference changes do not take effect until the page is refreshed.
This commit adds a warning to the preference screen if the user modified
a preference with this requirement.

Once the warning is triggered, it does not go away until the page is
refreshed, even if the user changes the preference back to the way it
was. Ideally we would detect this and hide the warning, but this was
more work so I didn't bother for this pass.